requirements.txt
asgiref==3.8.1
Django==5.0.7
django-autoslug==1.9.9
django-cors-headers==4.4.0
django-extensions-too==0.1.7
django-filter==24.2
django-froala-editor==4.3.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
